% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPALMOfromsinglecellmatrix.R
\name{createPALMOfromsinglecellmatrix}
\alias{createPALMOfromsinglecellmatrix}
\title{createPALMOfromsinglecellmatrix Function}
\usage{
createPALMOfromsinglecellmatrix(data, metadata, anndata = NULL)
}
\arguments{
\item{data}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample
column)}

\item{metadata}{Metadata associated with singe cell information. For example
rownames are unique cell_barcode and columns are information on each
cell_barcode like Sample (source of cell_barcode)}

\item{anndata}{Annotation dataframe. It consist of information such as
\emph{Sample} (sample name), \emph{PTID} (donor/participant), \emph{Time}
(longitudinal timepoints)}
}
\value{
PALMO object with scRNA
}
\description{
This function allows to create Seurat object from counts and metadata as
mentioned in
https://search.r-project.org/CRAN/refmans/SeuratObject/html/CreateSeuratObject.html.
The seurat object then stored in a newly created PALMO object.
}
\examples{
\dontrun{
palmo_obj=createPALMOfromsinglecellmatrix(counts, metadata, annotation)
}
}
\keyword{createPALMOfromsinglecellmatrix}
