% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvCalcBulkProfile.R
\name{cvCalcBulkProfile}
\alias{cvCalcBulkProfile}
\title{A cvCalcBulkProfile Function}
\usage{
cvCalcBulkProfile(ann, mat, fileName = NULL, filePATH = NULL)
}
\arguments{
\item{ann}{Annotation table. Table must consist column Sample (Participant
sample name), PTID (Participant), Time (longitudinal time points)}

\item{mat}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column)}

\item{fileName}{User-defined file name, Default outputFile}

\item{filePATH}{User-defined output directory PATH Default, current directory}
}
\description{
This function allows to calculate Intra-donor variations over longitudinal
timepoints. The coefficient of variation (CV) is calculated in Bulk
data without group information. CV calculated across samples. It requires
longitudinal data matrix/data frame and annotation file.
}
\keyword{cvCalcBulkProfile}
