% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvCalcSC.R
\name{cvCalcSC}
\alias{cvCalcSC}
\title{A cvCalcSC Function}
\usage{
cvCalcSC(
  ann,
  mat,
  meanThreshold = NULL,
  cvThreshold = NULL,
  housekeeping_genes = NULL,
  fileName = NULL,
  filePATH = NULL
)
}
\arguments{
\item{ann}{Annotation table. Table must consist column Sample (Participant
sample name), PTID (Participant), Time (longitudinal time points)}

\item{mat}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column)}

\item{meanThreshold}{Average expression threshold to filter lowly expressed
genes Default is 0.1 (log2 scale)}

\item{cvThreshold}{Coefficient of variation threshold to select variable and
stable genes Default is 10 for single cell RNA (100*SD/mean)}

\item{housekeeping_genes}{Optional list of housekeeping genes to focus on.
Default is ACTB, GAPDH}

\item{fileName}{User-defined file name, Default outputFile}

\item{filePATH}{User-defined output directory PATH Default, current directory}
}
\description{
This function allows to calculate Intra-donor variations over longitudinal
timepoints. The coefficient of variation is calculated in single cell
data. It requires longitudinal data matrix/data frame and annotation file.
}
\keyword{cvCalcSC}
