% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvSampleprofile.R
\name{cvSampleprofile}
\alias{cvSampleprofile}
\title{A cvSampleprofile Function}
\usage{
cvSampleprofile(
  mat,
  ann,
  meanThreshold = NULL,
  cvThreshold = NULL,
  fileName = NULL,
  filePATH = NULL
)
}
\arguments{
\item{mat}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column)}

\item{ann}{Annotation table. the must consist column Sample,
Participant sample name; PTID, participant; Time, longitudinal time frame; group,
name of the group; group_donor, combined string using group:Sample}

\item{meanThreshold}{Average expression threshold to filter lowly expressed
genes Default is 0.1 (log2 scale)}

\item{cvThreshold}{Coefficient of variation threshold to select variable and
stable genes Default is 10 for single cell RNA (100*SD/mean)}

\item{fileName}{User-defined file name, Default outputFile}

\item{filePATH}{User-defined output directory PATH Default, current directory}
}
\description{
This function allows to calculate Intra-donor variations over longitudinal
timepoints. The coefficient of variation is calculated in single cell
data. It requires longitudinal data matrix/data frame and annotation file.
}
\keyword{cvCalcSC}
