% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genecircosPlot.R
\name{genecircosPlot}
\alias{genecircosPlot}
\title{A genecircosPlot Function}
\usage{
genecircosPlot(
  data,
  geneList,
  groupColumn = NULL,
  group_oi = NULL,
  colorThreshold = NULL
)
}
\arguments{
\item{data}{Expression matrix or data frame. Rows represents gene/proteins
column represents group:donor (group and donor separated by :)}

\item{geneList}{Genes of interest to explore}

\item{groupColumn}{Default 1, use 2 when columns are donor:group format}

\item{group_oi}{Optional, User-defined groups to consider}

\item{colorThreshold}{User-defined color threshold in colorspace}
}
\description{
This function allows you to Circos Plot for gene list of interest by group
}
\examples{
##Circos Plot for genes expression in a group
#geneList <- c("IL32","CCL5","TCF7","IL7R","LEF1")
#res <- genecircosPlot(data=cv_res, geneList=geneList)
}
\keyword{genecircosPlot}
