% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqrBulk.R
\name{iqrBulk}
\alias{iqrBulk}
\title{A iqrBulk Function}
\usage{
iqrBulk(ann, mat, SD_threshold = NULL, fileName = NULL, filePATH = NULL)
}
\arguments{
\item{ann}{Annotation table. the must consist column Sample,
Participant sample name; PTID, participant; Time, longitudinal time frame}

\item{mat}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column)}

\item{SD_threshold}{Standard devaition limit to find outlliers (Eg. SD_threshold=
2, equals to Mean+/-2SD)}

\item{fileName}{User-defined file name, Default outputFile}

\item{filePATH}{User-defined output directory PATH Default, current directory}
}
\description{
This function allows you to perform outlier analysis using defining IQR in
bulk dataset. Outlier genes defined as exp > mean +/- 2SD.
}
\examples{
#filePATH <- getwd()
#IQR_res <- iqrBulk(ann=metadata, mat=datamatrix, fileName="olink", filePATH=filePATH)
}
\keyword{iqrBulk}
