% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqrSC.R
\name{iqrSC}
\alias{iqrSC}
\title{A iqrSC Function}
\usage{
iqrSC(
  ann,
  mat,
  groupName = NULL,
  SD_threshold = NULL,
  fileName = NULL,
  filePATH = NULL
)
}
\arguments{
\item{ann}{Annotation table. the must consist column Sample,
Participant sample name; PTID, participant; Time, longitudinal time frame}

\item{mat}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column)}

\item{groupName}{Group label used to calculate average gene expression
and check for outliers}

\item{SD_threshold}{Standard devaition limit to find outlliers (Eg. SD_threshold=
2, equals to Mean+/-2SD)}

\item{fileName}{User-defined file name, Default outputFile}

\item{filePATH}{User-defined output directory PATH Default, current directory}
}
\description{
This function allows you to perform outlier analysis on single cell based
average expression matrix by defining IQR. Outlier genes defined as exp >
mean +/- 2SD.
}
\examples{
#filePATH <- getwd()
#IQR_res <- iqrSC(ann=metadata, mat=datamatrix, fileName="olink", filePATH=filePATH)
}
\keyword{iqrSC}
