% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeVariance.R
\name{lmeVariance}
\alias{lmeVariance}
\title{A lmeVariance Function}
\usage{
lmeVariance(
  ann,
  mat,
  featureSet,
  meanThreshold = NULL,
  fileName = NULL,
  filePATH = NULL
)
}
\arguments{
\item{ann}{Annotation table. Table must consist column Sample (Participant
sample name), PTID (Participant), Time (longitudinal time points)}

\item{mat}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column)}

\item{featureSet}{Variance analysis carried out for the feature set provided such
as c("PTID", "Time", "Sex")}

\item{meanThreshold}{Average expression threshold to filter lowly expressed
genes/features Default is 0}

\item{fileName}{User-defined file name, Default outputFile}

\item{filePATH}{User-defined output directory PATH Default, current directory}
}
\description{
This function allows you to calculate inter-donor variation between participants
over longitudinal time points. It uses linear mixed model to calculate variance
contribution from each given feature list
}
\examples{
##Input Expression data
#filePATH <- getwd()
#lmem_res <- lmeVariance(ann=metadata, mat=datamatrix,
#featureSet=c("PTID", "Time", "Sex"),
#meanThreshold=0.1, fileName="RNA", filePATH=filePATH)
}
\keyword{lmeVariance}
