% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalDynamics.R
\name{longitudinalDynamics}
\alias{longitudinalDynamics}
\title{A longitudinalDynamics Function}
\usage{
longitudinalDynamics(
  metadata = NULL,
  data = NULL,
  datatype = NULL,
  omics = NULL,
  featureSet = NULL,
  meanThreshold = 1,
  cvThreshold = 5,
  NA_threshold = 0.4,
  column_sep = NULL,
  coding_genes = NULL,
  avgGroup = NULL,
  housekeeping_genes = c("ACTB", "GAPDH"),
  group_oi = NULL,
  nPC = 15,
  donorThreshold = NULL,
  groupThreshold = NULL,
  topFeatures = 25,
  method = "spearman",
  clusterBy = "donor",
  z_cutoff = 2,
  doOutlier = FALSE,
  fileName = NULL,
  outputDirectory = NULL
)
}
\arguments{
\item{metadata}{Annotation table. Table must consist column Sample (Participant
sample name), PTID (Participant), Time (longitudinal time points)}

\item{data}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column).
For single cell, Single cell RNA Seurat object, if datatype is single cell RNA
and Single cell ATAC genescore matrix or data frame}

\item{datatype}{Data input can be bulk or singlecell}

\item{omics}{User defined name like RNA, ATAC, Proteomics, FLOW}

\item{featureSet}{Variance analysis carried out on the featureSet provided such
as c("PTID", "Time", "Sex")}

\item{meanThreshold}{Average expression threshold to filter lowly expressed
genes Default is 0.1 (log2 scale)}

\item{cvThreshold}{Coefficient of variation threshold to select variable and
stable genes Default is 10 for single cell RNA (100*SD/mean)}

\item{NA_threshold}{Number of NAs in data (numeric value or NULL). Default,
40\% * number of columns.}

\item{column_sep}{Separator of "PTID" and "Time" in "Sample" column of
Annotation table like column_sep="W" for PTID1W1, column_sep=":" for
PTID1W1:Tcell}

\item{coding_genes}{Selecting protein coding/user-defined gene list only}

\item{avgGroup}{Group label to be used to calculate average gene expression by
group label}

\item{housekeeping_genes}{Optional list of housekeeping genes to focus on
Default is NULL}

\item{group_oi}{Group of interest to focus on, Default is NULL}

\item{nPC}{Number of PCAs to be used for UMAP, Default is 15}

\item{donorThreshold}{Donor threshold number to be used, Default is number of
participants}

\item{groupThreshold}{Group label threshold number to be used, Default is
(number of participants x group labels)/2}

\item{topFeatures}{Number of features to be selected from each group, Default
is 25}

\item{method}{Sample correlation analysis ("pearson", "spearman"). Default is
"spearman"}

\item{clusterBy}{for sample correlation cluster columns by ("donor", "group")}

\item{z_cutoff}{|Z| cutoff threshold to find potential outliers (Eg. z_cutoff=
2, equals to Mean/SD 2)}

\item{doOutlier}{To perform outlier analysis (TRUE or FALSE). Default FALSE}

\item{fileName}{User defined filename}

\item{outputDirectory}{User-defined output directory Default, output}
}
\description{
This function allows you to perform analysis of longitudinal dataset.
It requires longitudinal data matrix/data frame and annotation file.
}
\keyword{longitudinalDynamics}
