% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalDynamics_bulk_analysis.R
\name{longitudinalDynamics_bulk_analysis}
\alias{longitudinalDynamics_bulk_analysis}
\title{A longitudinalDynamics_bulk_analysis Function}
\usage{
longitudinalDynamics_bulk_analysis(
  metadata = NULL,
  datamatrix = NULL,
  datatype = NULL,
  omics = NULL,
  fileName = NULL,
  filePATH = NULL,
  features = NULL,
  method = NULL,
  meanThreshold = NULL,
  cvThreshold = NULL
)
}
\arguments{
\item{metadata}{Annotation table must consist column Sample,
Participant sample name; PTID, participant; Time, longitudinal time frame}

\item{datamatrix}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column)}

\item{datatype}{Data input can be bulk or singlecell}

\item{omics}{User defined name like RNA, ATAC, Proteomics, FLOW}

\item{fileName}{User defined filename}

\item{filePATH}{User-defined output directory path}

\item{features}{Variance analysis carried out for the features provided such
as c("PTID", "Time", "Sex")}

\item{method}{Type of variations to be analyzed, c("intra-donor", "inter-donor",
"iqr"), Default is all}

\item{meanThreshold}{Average expression threshold to filter lowly expressed
genes Default is 0.1 (log2 scale)}

\item{cvThreshold}{Coefficient of variation threshold to select variable and
stable genes Default is 10 for single cell RNA (100*SD/mean)}
}
\description{
This function allows you to perform analysis of longitudinal data from bulk
dataset. It requires longitudinal bulk data matrix/data frame at sample level
and annotation file.
}
\examples{
##Bulk data
#longitudinalDynamics(metadata=ann, datamatrix=data, datatype="bulk", omics="Protein",
#fileName="olink", features=c("PTID", "Time", "Sex"), meanThreshold=1, cvThreshold=5,
#outputDirectory="output")
}
\keyword{longitudinalDynamics_bulk_analysis}
