% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalDynamics_scatac_analysis.R
\name{longitudinalDynamics_scatac_analysis}
\alias{longitudinalDynamics_scatac_analysis}
\title{A longitudinalDynamics_scatac_analysis Function}
\usage{
longitudinalDynamics_scatac_analysis(
  metadata = NULL,
  genescore = NULL,
  datatype = NULL,
  omics = NULL,
  fileName = NULL,
  filePATH = NULL,
  features = NULL,
  meanThreshold = 0.1,
  cvThreshold = 10,
  coding_genes = NULL,
  avgGroup = NULL,
  housekeeping_genes = NULL,
  group_oi = NULL,
  nPC = NULL,
  method = NULL,
  donorThreshold = NULL,
  groupThreshold = NULL,
  topFeatures = 25
)
}
\arguments{
\item{metadata}{Annotation table must consist column Sample,
Participant sample name; PTID, participant; Time, longitudinal time frame}

\item{genescore}{Genescore Expression matrix or data frame derived from scATAC.
Rows represents genes and column represents participant samples (same as
annotation table Sample column)}

\item{datatype}{Data input can be bulk or singlecell}

\item{omics}{User defined name like RNA, ATAC, Proteomics, FLOW}

\item{fileName}{User defined filename}

\item{filePATH}{User-defined output directory path}

\item{features}{Variance analysis carried out for the features provided such
as c("PTID", "Time", "Sex")}

\item{meanThreshold}{Average expression threshold to filter lowly expressed
genes Default is 0.1 (log2 scale)}

\item{cvThreshold}{Coefficient of variation threshold to select variable and
stable genes Default is 10 for single cell RNA (100*SD/mean)}

\item{coding_genes}{Selecting protein coding/user-defined gene list only}

\item{avgGroup}{Group label to be used to calculate average gene expression by
group label}

\item{housekeeping_genes}{Optional list of housekeeping genes to focus on
Default is NULL}

\item{group_oi}{Group of interest to focus on, Default is NULL}

\item{nPC}{Number of PCAs to be used for UMAP, Default is 30}

\item{method}{Type of variations to be analyzed, c("intra-donor", "inter-donor",
"iqr"), Default is all}

\item{donorThreshold}{Donor threshold number to be used, Default is number of
participants}

\item{groupThreshold}{Group label threshold number to be used, Default is
(number of participants x group labels)/2}

\item{topFeatures}{Number of features to be selected from each group, Default
is 25}
}
\description{
This function allows you to perform analysis of longitudinal dataset from
single cell atac based genescore matrix.
}
\examples{
##Single cell ATAC data
#longitudinalDynamics(metadata=ann, atacObj=atacObj, 
#datatype="singlecell", omics="atac",fileName="scatac",
#features=c("PTID", "Time", "group"),
#meanThreshold=0.1, cvThreshold=10,
#housekeeping_genes=NULL, group_oi=NULL,
#nPC=30, donorThreshold=4, groupThreshold=40, topFeatures=25,
#method=c("intra-donor","inter-donor","iqr"),
#outputDirectory="output")
}
\keyword{longitudinalDynamics_scatac_analysis}
