% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalDynamics_scrna_analysis.R
\name{longitudinalDynamics_scrna_analysis}
\alias{longitudinalDynamics_scrna_analysis}
\title{A longitudinalDynamics_scrna_analysis Function}
\usage{
longitudinalDynamics_scrna_analysis(
  metadata = NULL,
  dataObj = NULL,
  datatype = "singlecell",
  omics = "scRNA",
  fileName = NULL,
  filePATH = NULL,
  features = NULL,
  meanThreshold = 0.1,
  cvThreshold = 10,
  coding_genes = NULL,
  avgGroup = NULL,
  housekeeping_genes = NULL,
  group_oi = NULL,
  nPC = NULL,
  method = NULL,
  donorThreshold = NULL,
  groupThreshold = NULL,
  topFeatures = 25
)
}
\arguments{
\item{metadata}{Annotation table must consist column Sample,
Participant sample name; PTID, participant; Time, longitudinal time frame}

\item{dataObj}{Single cell Expression data. The scRNA data should be seurat
object. To get more information visit https://cran.r-project.org/web/packages/SeuratObject/SeuratObject.pdf}

\item{datatype}{Default singlecell}

\item{omics}{Default scRNA}

\item{fileName}{User defined filename}

\item{filePATH}{User-defined output directory path}

\item{features}{Variance analysis carried out for the features provided such
as c("PTID", "Time", "Sex")}

\item{meanThreshold}{Average expression threshold to filter lowly expressed
genes Default is 0.1 (log2 scale)}

\item{cvThreshold}{Coefficient of variation threshold to select variable and
stable genes Default is 10 for single cell RNA (100*SD/mean)}

\item{coding_genes}{Selecting protein coding/user-defined gene list only}

\item{avgGroup}{Group label to be used to calculate average gene expression by
group label}

\item{housekeeping_genes}{Optional list of housekeeping genes to focus on
Default is NULL}

\item{group_oi}{Group of interest to focus on, Default is NULL}

\item{nPC}{Number of PCAs to be used for UMAP, Default is 30}

\item{method}{Type of variations to be analyzed, c("intra-donor", "inter-donor",
"iqr"), Default is all}

\item{donorThreshold}{Donor threshold number to be used, Default is number of
participants}

\item{groupThreshold}{Group label threshold number to be used, Default is
(number of participants x group labels)/2}

\item{topFeatures}{Number of features to be selected from each group, Default
is 25}
}
\description{
This function allows you to perform analysis of longitudinal dataset from
single cell RNA.
}
\examples{
##Single cell RNA data
#load("data/pbmc1-subset.Rda")
#load("data/data_Annotation.Rda")
#metadata=ann
#long_res <- longitudinalDynamics(metadata=ann, scrnaObj=pbmc,
#datatype="singlecell", omics="rna",
#fileName="scrna", features=c("PTID", "Time"),
#meanThreshold=0.1, cvThreshold=10,
#coding_genes=FALSE, avgGroup="celltype",
#housekeeping_genes=housekeeping_genes,
#nPC=15,
#donorThreshold=4, groupThreshold=40, topFeatures=25,
#method=c("intra-donor","inter-donor","iqr"),
#outputDirectory="output")
}
\keyword{longitudinalDynamics_scrna_analysis}
