% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePALMOdata.R
\name{mergePALMOdata}
\alias{mergePALMOdata}
\title{mergePALMOdata Function}
\usage{
mergePALMOdata(data_object, datatype)
}
\arguments{
\item{data_object}{Input \emph{PALMO} S4 object. It contains annotation
information and expression data from Bulk or single cell data.}

\item{datatype}{Input data type 'bulk' or 'singlecell'}
}
\value{
PALMO object with merged annotation and data matrix
}
\description{
This function allows to merge expression data from bulk or single cell data
with annotation file provided by user. It will remove the annotations with
missing information from Sample name, donor/participant and time variable.
}
\examples{
\dontrun{
palmo_obj <- mergePALMOdata(data_object=palmo_obj)
}
}
\keyword{mergePALMOdata}
