% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetectP.R
\name{outlierDetectP}
\alias{outlierDetectP}
\title{A outlierDetectP Function}
\usage{
outlierDetectP(outlier_events, z_cutoff = 2, nGenes, groupBy = "PTID")
}
\arguments{
\item{outlier_events}{Identified outlier events}

\item{z_cutoff}{|Z| cutoff threshold to find potential outliers (Eg. z_cutoff=
2, equals to Mean/SD 2)}

\item{nGenes}{Number of background genes/features}

\item{groupBy}{Column name to use for groupwise outlier analysis default is
PTID (patient id)}
}
\description{
This function allows to identify significant abnormal event identified
from outlier analysis.
}
\examples{
#outlierDetectP(outlier_events, z_cutoff=2, nGenes)
}
\keyword{outlierDetectP}
