% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetectP.R
\name{p_value_for_event}
\alias{p_value_for_event}
\title{p_value_for_event Function}
\usage{
p_value_for_event(events, tries, rate)
}
\arguments{
\item{events}{Identified outlier events}

\item{tries}{Number of background genes/features}

\item{rate}{probability distribution}
}
\value{
outlier event p value
}
\description{
This function allows to calculate p value for identified outlier significant
abnormal events
}
\examples{
\dontrun{
p_value_for_event(events, tries, rate)
}
}
\keyword{p_value_for_event}
