% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_correlation.R
\name{sample_correlation}
\alias{sample_correlation}
\title{A sample_correlation Function}
\usage{
sample_correlation(
  data,
  column_sep = ":",
  method = "spearman",
  groupColumn = 2,
  clusterBy = "donor",
  max = 0.9,
  column_names_fontsize = 4,
  row_names_fontsize = 4,
  row_title_fontsize = 6,
  column_title_fontsize = 6,
  plotHeight = 20,
  fileName = NULL,
  filePATH = NULL
)
}
\arguments{
\item{data}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (if celltype with in donor then sample:
celltype, separated by :)}

\item{column_sep}{Sample and celltype seperator like (:)}

\item{method}{Correlation method "pearson" or "spearman"}

\item{groupColumn}{Data column names consists group (Donor-group) at 2nd place
or 1st place(like PTIDxGroupX, 2 or GroupXPTIDx, 1)}

\item{clusterBy}{Cluster correlation result by "donor" or "group". Default
donor}

\item{max}{Maximum color limit (Default, 0.9 correlation)}

\item{column_names_fontsize}{Font size of the column names, default 4}

\item{row_names_fontsize}{Font size of the row names, default 4}

\item{row_title_fontsize}{Font size of the row title, default 6}

\item{column_title_fontsize}{Font size of the column title, default 6}

\item{plotHeight}{Height of the plot (in), deafult 20in}

\item{fileName}{User-defined file name, Default outputFile}

\item{filePATH}{User-defined output directory PATH Default, current directory}
}
\description{
This function allows to perform sample correlation (by group like
celltype, ot by donor).
}
\examples{
#res <- sample_correlation(data=datamatrix, column_sep=":", method="spearman")
}
\keyword{sample_correlation}
