% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatac_archr_genescore.R
\name{scatac_archr_genescore}
\alias{scatac_archr_genescore}
\title{A scatac_archr_genescore Function}
\usage{
scatac_archr_genescore(ArchRProj, groupBy)
}
\arguments{
\item{ArchRProj}{archR scATAC object for input single cell ATAC longitudinal data}

\item{groupBy}{Group label to be used to calculate average gene expression by
group label, Eg. "celltype"}
}
\description{
This function allows you to calculate genescore matrix from scATAC archR object.
This function requires archR package installed and scATAC object created.
}
\examples{
##Input scATAC data
#genescore <- scatac_archr_genescore(ArchRProj=proj, groupBy="celltype")
}
\keyword{scatac_archr_genescore}
