% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclongitudinalDGE.R
\name{sclongitudinalDGE}
\alias{sclongitudinalDGE}
\title{A sclongitudinalDGE Function}
\usage{
sclongitudinalDGE(
  ann,
  dataObj,
  scassay = "RNA",
  celltypecol,
  mincellsexpressed = 0.1,
  removelnc = "TRUE",
  adjfac = "none",
  baseline = NULL,
  plotWidth = 10,
  plotHeight = 10,
  fileName = NULL,
  filePATH = NULL
)
}
\arguments{
\item{ann}{Annotation dataframe. Table must consist column Sample (Participant
sample name), PTID (Participant), Time (longitudinal time points)}

\item{dataObj}{Single cell RNA seurat object. Seurat object should have column
name Sample (same as annotation table Sample column)}

\item{scassay}{Single cell assay from scRNA seurat object (Default "RNA")}

\item{celltypecol}{Column of interest such as celltype to analyze DEGs in
participant over time}

\item{mincellsexpressed}{Average expression threshold to filter lowly expressed
genes/features Default is 0.1}

\item{removelnc}{Remove lincRNAs, mitochondrial and ribosomal genes from
analysis incldes (^RP|^MT-|^LINC|orf) (TRUE/FALSE). Default is TRUE}

\item{adjfac}{Factors to be adjusted for such as batch, sex}

\item{baseline}{Donors (PTID) to be considered as baseline. Deafult NULL}

\item{plotWidth}{User-defined plot width, Default 10 in}

\item{plotHeight}{User-defined plot height, Default 10 in}

\item{fileName}{User-defined file name, Default outputFile}

\item{filePATH}{User-defined output directory PATH Default, current directory}
}
\description{
This function allows you to calculate differential expressed genes in the
direction of given time points (if timepoints>3 otherwise DEGs between two
timepoints). A hurdle model was fit to each participant independently in
order to identify participant-specific longitudinal transcriptomic changes.
Genes that were expressed in at least 10\% of cells per participant were
considered for this analysis. The models were fit on the input normalized
data, modeling the timepoints as a continuous variable within each cell type
and adjusting for the batch only if any timepoints from the same participant
were run across multiple batches.
}
\examples{
##Input scRNA data and annotation file
#DEGres <- sclongitudinalDGE(ann=metadata, dataObj=pbmc, scassay="RNA", celltypecol="celltype")
}
\keyword{sclongitudinalDGE}
