% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genePlot.R
\name{genePlot}
\alias{genePlot}
\title{A genePlot Function}
\usage{
genePlot(ann = NULL, data = NULL, geneName = NULL, groupName = NULL)
}
\arguments{
\item{ann}{Optional. Annotation table used in case of count/genescore
matrix. It must consist column Sample, Participant sample name; PTID,
participant; Time, longitudinal time frame; group,
name of the group; group_donor, combined string using group:Sample}

\item{data}{Average Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples with group (optional).}

\item{geneName}{User-defined gene name}

\item{groupName}{User-defined group name column from annotation table}
}
\description{
This function allows you to perform UMAP visualization of gene of interest list.
}
\examples{
#plot <- genePlot(ann=annotation, data=ExpressionData, geneName="FOLR3", groupName="Time")
}
\keyword{genePlot}
