% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalDynamics.R
\name{longitudinalDynamics}
\alias{longitudinalDynamics}
\title{A longitudinalDynamics Function}
\usage{
longitudinalDynamics(
  metadata = NULL,
  datamatrix = NULL,
  scrnaObj = NULL,
  atacObj = NULL,
  datatype = NULL,
  omics = NULL,
  fileName,
  features = NULL,
  meanThreshold = 1,
  cvThreshold = 5,
  coding_genes = NULL,
  avgGroup = NULL,
  housekeeping_genes = NULL,
  group_oi = NULL,
  nPC = 30,
  donorThreshold = NULL,
  groupThreshold = NULL,
  topFeatures = 25,
  method = NULL,
  outputDirectory = NULL
)
}
\arguments{
\item{metadata}{Annotation table must consist column Sample,
Participant sample name; PTID, participant; Time, longitudinal time frame}

\item{datamatrix}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column)}

\item{scrnaObj}{Single cell RNA Seurat object, if datatype is single cell RNA}

\item{atacObj}{Single cell ATAC genematrix or data frame, if datatype is
single cell ATAC}

\item{datatype}{Data input can be bulk or singlecell}

\item{omics}{User defined name like RNA, ATAC, Proteomics, FLOW}

\item{fileName}{User defined filename}

\item{features}{Variance analysis carried out for the features provided such
as c("PTID", "Time", "Sex")}

\item{meanThreshold}{Average expression threshold to filter lowly expressed
genes Default is 0.1 (log2 scale)}

\item{cvThreshold}{Coefficient of variation threshold to select variable and
stable genes Default is 10 for single cell RNA (100*SD/mean)}

\item{coding_genes}{Selecting protein coding/user-defined gene list only}

\item{avgGroup}{Group label to be used to calculate average gene expression by
group label}

\item{housekeeping_genes}{Optional list of housekeeping genes to focus on
Default is NULL}

\item{group_oi}{Group of interest to focus on, Default is NULL}

\item{nPC}{Number of PCAs to be used for UMAP, Default is 30}

\item{donorThreshold}{Donor threshold number to be used, Default is number of
participants}

\item{groupThreshold}{Group label threshold number to be used, Default is
(number of participants x group labels)/2}

\item{topFeatures}{Number of features to be selected from each group, Default
is 25}

\item{method}{Type of variations to be analyzed, c("intra-donor", "inter-donor",
"iqr"), Default is all}

\item{outputDirectory}{User-defined output directory Default, output}
}
\description{
This function allows you to perform analysis of longitudinal dataset.
It requires longitudinal data matrix/data frame and annotation file.
}
\examples{
##Bulk data
#load("data/data_Annotation.Rda")
#load("data/Olink_NPX_log2_Protein.Rda")
#longitudinalDynamics(metadata=ann, datamatrix=data, datatype="bulk", omics="Protein",
#fileName="olink", features=c("PTID", "Time", "Sex"), meanThreshold=1, cvThreshold=5,
#outputDirectory="output")

##Single cell RNA data
#load("data/pbmc1-subset.Rda")
#load("data/data_Annotation.Rda")
#metadata=ann
#long_res <- longitudinalDynamics(metadata=ann, scrnaObj=pbmc,
#datatype="singlecell", omics="rna",
#fileName="scrna", features=c("PTID", "Time"),
#meanThreshold=0.1, cvThreshold=10,
#coding_genes=FALSE, avgGroup="celltype",
#housekeeping_genes=housekeeping_genes,
#nPC=15,
#donorThreshold=4, groupThreshold=40, topFeatures=25,
#method=c("intra-donor","inter-donor","iqr"),
#outputDirectory="output")
}
\keyword{longitudinalDynamics}
