% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodalView.R
\name{multimodalView}
\alias{multimodalView}
\title{A multimodalView Function}
\usage{
multimodalView(
  modality1,
  modality2,
  groupBy = NULL,
  geneList,
  colorThreshold = 10,
  groupColumn = NULL,
  plotHeight = 10,
  fileName = NULL,
  filePATH = NULL
)
}
\arguments{
\item{modality1}{Variation or Expression matrix/data frame. Rows represents
gene/proteins column represents group:donor (group and donor separated by :)}

\item{modality2}{Variation or Expression matrix/data frame. Rows represents
gene/proteins column represents group:donor (group and donor separated by :)}

\item{groupBy}{Optional, User-defined groups to consider and order}

\item{geneList}{Genes of interest to explore}

\item{colorThreshold}{User-defined color threshold in colorspace}

\item{groupColumn}{Default 1, use 2 when columns are donor:group format}

\item{plotHeight}{User-defined Plot size (in)}

\item{fileName}{User defined filename}

\item{filePATH}{User-defined output directory path to save result}
}
\description{
This function allows you to vizualize the multimodal view genes of interest
by celltypes/ groups defined by use
}
\examples{
##Circos Plot for genes expression in a group
#geneList <- c("HLA-A","HLA-B","HLA-C","HLA-DRA","HLA-DPA1","HLA-DRB1")
#multimodalView(modality1=scrna_cv_res, modality2=scatac_cv_res, geneList)
}
\keyword{multimodalView}
