% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetect.R
\name{outlierDetect}
\alias{outlierDetect}
\title{A outlierDetect Function}
\usage{
outlierDetect(
  ann,
  mat,
  z_cutoff = 2,
  plotWidth = 10,
  plotHeight = 5,
  groupBy = FALSE,
  fileName = NULL,
  filePATH = NULL
)
}
\arguments{
\item{ann}{Annotation table. Table must consist column Sample (Participant
sample name), PTID (Participant), Time (longitudinal time points)}

\item{mat}{Expression matrix or data frame. Rows represents gene/proteins
column represents participant samples (same as annotation table Sample column)}

\item{z_cutoff}{|Z| cutoff threshold to find potential outliers (Eg. z_cutoff=
2, equals to Mean/SD 2)}

\item{plotWidth}{User-defined plot width, Default 10 in}

\item{plotHeight}{User-defined plot height, Default 5 in}

\item{groupBy}{Include groupwise outlier analysis (TRUE or FALSE). Column
used for analyis is Sample_group}

\item{fileName}{User-defined file name, Default outputFile}

\item{filePATH}{User-defined output directory PATH Default, current directory}
}
\description{
This function allows you to perform outlier analysis on bulk data by
calculating z-score. Outlier genes defined as mean/SD = |Z| > z_cutoff.
}
\examples{
#filePATH <- getwd()
#outlier_res <- outlierDetect(ann=metadata, mat=datamatrix)
}
\keyword{outlierDetect}
