% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetectP.R
\name{outlierDetectP}
\alias{outlierDetectP}
\title{outlierDetectP Function}
\usage{
outlierDetectP(
  outlier_events,
  z_cutoff = 2,
  nGenes,
  group_by = "PTID",
  alternative = "two.sided"
)
}
\arguments{
\item{outlier_events}{Identified outlier events}

\item{z_cutoff}{|Z| cutoff threshold to find potential outliers (Eg.
z_cutoff= 2, equals to Mean/SD 2)}

\item{nGenes}{Number of background genes/features}

\item{group_by}{Column name to use for groupwise outlier analysis. Default is
PTID (donor or participant id).}

\item{alternative}{alternative hypothesis, must be one of "one.sided"
or "two.sided" (default)}
}
\value{
PALMO object with outlier event p value dataframe

PALMO object with outlier event p value
}
\description{
This function allows to identify significant abnormal event identified
from outlier analysis.
}
\examples{
\dontrun{
outlierDetectP(outlier_events=outlier_res, z_cutoff=2, nGenes=1043)
}
}
\keyword{outlierDetectP}
